Refactor Viewer to accept pluggable content functions
Prepare the viewer for reuse by the VS Code extension webview:

- Change ViewerProps from rawBaseUrl to getTextContent + getAssetURL
  functions, so callers provide their own content loading strategy
- Extract URL-based fetch logic into createUrlContentFns() helper
- Move @tailwindcss/vite and tailwindcss to devDependencies (they're
  build-time tools, not runtime dependencies)

The viewer app passes createUrlContentFns(rawBaseUrl) to build the
functions from a GitHub raw URL. The extension webview will pass
workspace-based implementations instead.

No functional change — viewer builds, all 60 tests pass.
