+++ apps/viewer/package.json
@@ -12,18 +12,18 @@
     "test:watch": "vitest"
   },
   "dependencies": {
-    "@tailwindcss/vite": "^4.2.2",
     "js-yaml": "^4.1.1",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
-    "stagebook": "*",
-    "tailwindcss": "^4.2.2"
+    "stagebook": "*"
   },
   "devDependencies": {
+    "@tailwindcss/vite": "^4.2.2",
     "@types/js-yaml": "^4.0.9",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^4.5.2",
+    "tailwindcss": "^4.2.2",
     "typescript": "^5.5.0",
     "vite": "^6.3.3",
     "vitest": "^4.1.4"
